dom/media/tests/mochitest/TextTrack_active_cues_null.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=833386
-->
<head>
  <title>Test for Bug 833386</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=833386">Mozilla Bug 833386</a>
<p id="display"></p>
<div id="content" style="display: none">
  <video id="video_1" src="demo.webm">
    <track kind="subtitles" src="overlap.vtt" default />
  </video>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
  SpecialPowers.setBoolPref("media.webvtt.enabled", true);
  SimpleTest.waitForExplicitFinish();
  
  var textTrack;

  function start() {
    var video = document.getElementById('video_1');
    textTrack = video.textTracks[0];
    textTrack.mode = 'disabled';

    is(textTrack.activeCues, null);
    SimpleTest.finish();
  }
  addLoadEvent(start);
</script>
</pre>
</body>
</html>
